+++ RELEASE_NOTES.rst
@@ -9,6 +9,7 @@ ARTIQ-7
 Highlights:
 * WRPLL
 * ``get()``, ``get_mu()``, ``get_att()``, and ``get_att_mu()`` functions added for AD9910 and AD9912
+* ``EnumerationValue`` with one choice and no default will default to the single available choice
 
 Breaking changes: